Require unauthenticated GET for renewalInfo requests
Do not allow POST-as-GET for accessing renewalInfo resources; instead
require that servers only expose the resource via GET, and that clients
only query the resource via GET.

This reduces the possible implementation space for clients, simplifying
the work for client maintainers. It also ensures that the renewalInfo
responses can be easily cached, allowing server operators to minimize
the burden of supporting this draft.