Construct renewalInfo URLs from directory+fingerprint (#2)
Instead of providing renewalInfo URLs inside order objects, requiring
the ACME client to persist the URL in order to query it on a regular
basis, allow renewalInfo URLs to be constructable given only existing
client configuration (i.e. the URL of the directory) and the certificate
itself.

This allows stateless clients to not need to persist additional data,
and allows external monitoring tools to poll for renewal information
without being privy to order objects, at the cost of requiring two API
calls (one for the directory, one for the renewalInfo) each time a
client checks in.